fix: keterangan_konversi_nilai
- change int rentang_atas, bawah -> float
- keterangan_konversi_nilai -> 83,33 == null

+++ database/migrations/2024_03_20_000005_tb_rentang_nilai.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('keterangan_rentang_nilai');
             $table->string('keterangan');
             
-            $table->integer('rentang_atas');
-            $table->integer('rentang_bawah');
+            $table->float('rentang_atas', 8, 1);
+            $table->float('rentang_bawah', 8, 1);
 
             $table->integer('created_by')->nullable(); 
             $table->integer('updated_by')->nullable();
